Porting https://github.com/Activiti/Activiti/tree/entity-cache-proposal branch to the latest state of v6: introducing two new experimental flags to 1) cache execution trees during the lifetime of a command and 2) keep a count on executions to avoid deletes. This commit contains the basics to get the -Pcheck running and the optimization config working. However, still needs a general rollout. Also, timers/suspended/DL jobs not yet done.

+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/persistence/entity/ExecutionEntityManagerImpl.java
@@ -32,6 +32,7 @@ import org.activiti.engine.impl.ProcessInstanceQueryImpl;
 import org.activiti.engine.impl.cfg.ProcessEngineConfigurationImpl;
 import org.activiti.engine.impl.context.Context;
 import org.activiti.engine.impl.identity.Authentication;
+import org.activiti.engine.impl.persistence.CountingExecutionEntity;
 import org.activiti.engine.impl.persistence.entity.data.DataManager;
 import org.activiti.engine.impl.persistence.entity.data.ExecutionDataManager;
 import org.activiti.engine.runtime.Execution;
@@ -50,9 +51,12 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
   
   protected ExecutionDataManager executionDataManager;
   
+  protected boolean enableExecutionRelationshipCounts;
+  
   public ExecutionEntityManagerImpl(ProcessEngineConfigurationImpl processEngineConfiguration, ExecutionDataManager executionDataManager) {
     super(processEngineConfiguration);
     this.executionDataManager = executionDataManager;
+    this.enableExecutionRelationshipCounts = processEngineConfiguration.isEnableExecutionRelationshipCounts();
   }
 
   @Override
@@ -64,8 +68,7 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
   
   @Override
   public void delete(ExecutionEntity entity) {
-    super.delete(entity);
-    entity.setDeleted(true);
+    delete(entity, true);
   }
   
   @Override
@@ -118,8 +121,19 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
   
   @Override
   public ExecutionEntity findByRootProcessInstanceId(String rootProcessInstanceId) {
-    ExecutionEntity rootExecution = null;
     List<ExecutionEntity> executions = executionDataManager.findExecutionsByRootProcessInstanceId(rootProcessInstanceId);
+    return processExecutionTree(rootProcessInstanceId, executions);
+    
+  }
+  
+  /**
+   * Processes a collection of {@link ExecutionEntity} instances, which form on execution tree.
+   * All the executions share the same rootProcessInstanceId (which is provided).
+   * The return value will be the root {@link ExecutionEntity} instance, with all child {@link ExecutionEntity}
+   * instances populated and set using the {@link ExecutionEntity} instances from the provided collections
+   */
+  protected ExecutionEntity processExecutionTree(String rootProcessInstanceId, List<ExecutionEntity> executions) {
+    ExecutionEntity rootExecution = null;
     
     // Collect executions
     Map<String, ExecutionEntity> executionMap = new HashMap<String, ExecutionEntity>(executions.size());
@@ -132,6 +146,11 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
     
     // Set relationships
     for (ExecutionEntity executionEntity : executions) {
+      
+      // Root process instance relationship
+      if (executionEntity.getRootProcessInstanceId() != null) {
+        executionEntity.setRootProcessInstance(executionMap.get(executionEntity.getRootProcessInstanceId()));
+      }
       
       // Process instance relationship
       if (executionEntity.getProcessInstanceId() != null) {
@@ -153,9 +172,7 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
       }
       
     }
-    
     return rootExecution;
-    
   }
   
   @Override
@@ -487,7 +504,9 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
     executionEntity.setEnded(true);
     executionEntity.setActive(false);
     
-    if (executionEntity.getId().equals(executionEntity.getProcessInstanceId())) {
+    if (executionEntity.getId().equals(executionEntity.getProcessInstanceId())
+        && (!enableExecutionRelationshipCounts 
+            || (enableExecutionRelationshipCounts && ((CountingExecutionEntity) executionEntity).getIdentityLinkCount() > 0))) {
       IdentityLinkEntityManager identityLinkEntityManager = getIdentityLinkEntityManager();
       Collection<IdentityLinkEntity> identityLinks = identityLinkEntityManager.findIdentityLinksByProcessInstanceId(executionEntity.getProcessInstanceId());
       for (IdentityLinkEntity identityLink : identityLinks) {
@@ -496,20 +515,28 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
     }
 
     // Get variables related to execution and delete them
-    VariableInstanceEntityManager variableInstanceEntityManager = getVariableInstanceEntityManager();
-    Collection<VariableInstanceEntity> executionVariables = variableInstanceEntityManager.findVariableInstancesByExecutionId(executionEntity.getId());
-    for (VariableInstanceEntity variableInstanceEntity : executionVariables) {
-      variableInstanceEntityManager.delete(variableInstanceEntity);
-      if (variableInstanceEntity.getByteArrayRef() != null && variableInstanceEntity.getByteArrayRef().getId() != null) {
-        getByteArrayEntityManager().deleteByteArrayById(variableInstanceEntity.getByteArrayRef().getId());
+    if (!enableExecutionRelationshipCounts || 
+        (enableExecutionRelationshipCounts && ((CountingExecutionEntity) executionEntity).getVariableCount() > 0)) {
+      Collection<VariableInstance> executionVariables = executionEntity.getVariableInstancesLocal().values();
+      for (VariableInstance variableInstance : executionVariables) {
+        VariableInstanceEntity variableInstanceEntity = (VariableInstanceEntity) variableInstance;
+        
+        VariableInstanceEntityManager variableInstanceEntityManager = getVariableInstanceEntityManager();
+        variableInstanceEntityManager.delete(variableInstanceEntity);
+        if (variableInstanceEntity.getByteArrayRef() != null && variableInstanceEntity.getByteArrayRef().getId() != null) {
+          getByteArrayEntityManager().deleteByteArrayById(variableInstanceEntity.getByteArrayRef().getId());
+        }
       }
     }
 
     // Delete current user tasks
-    TaskEntityManager taskEntityManager = getTaskEntityManager();
-    Collection<TaskEntity> tasksForExecution = taskEntityManager.findTasksByExecutionId(executionEntity.getId());
-    for (TaskEntity taskEntity : tasksForExecution) {
-      taskEntityManager.deleteTask(taskEntity, deleteReason, false, cancel);
+    if (!enableExecutionRelationshipCounts || 
+        (enableExecutionRelationshipCounts && ((CountingExecutionEntity) executionEntity).getTaskCount() > 0)) {
+      TaskEntityManager taskEntityManager = getTaskEntityManager();
+      Collection<TaskEntity> tasksForExecution = taskEntityManager.findTasksByExecutionId(executionEntity.getId());
+      for (TaskEntity taskEntity : tasksForExecution) {
+        taskEntityManager.deleteTask(taskEntity, deleteReason, false, cancel);
+      }
     }
 
     // Delete jobs
@@ -522,12 +549,15 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
       }
     }
     
-    JobEntityManager jobEntityManager = getJobEntityManager();
-    Collection<JobEntity> jobsForExecution = jobEntityManager.findJobsByExecutionId(executionEntity.getId());
-    for (JobEntity job : jobsForExecution) {
-      jobEntityManager.delete(job);
-      if (getEventDispatcher().isEnabled()) {
-        getEventDispatcher().dispatchEvent(ActivitiEventBuilder.createEntityEvent(ActivitiEventType.JOB_CANCELED, job));
+    if (!enableExecutionRelationshipCounts 
+        || (enableExecutionRelationshipCounts && ((CountingExecutionEntity) executionEntity).getJobCount() > 0)) {
+      JobEntityManager jobEntityManager = getJobEntityManager();
+      Collection<JobEntity> jobsForExecution = jobEntityManager.findJobsByExecutionId(executionEntity.getId());
+      for (JobEntity job : jobsForExecution) {
+        getJobEntityManager().delete(job);
+        if (getEventDispatcher().isEnabled()) {
+          getEventDispatcher().dispatchEvent(ActivitiEventBuilder.createEntityEvent(ActivitiEventType.JOB_CANCELED, job));
+        }
       }
     }
     
@@ -550,10 +580,13 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
     }
 
     // Delete event subscriptions
-    EventSubscriptionEntityManager eventSubscriptionEntityManager = getEventSubscriptionEntityManager();
-    List<EventSubscriptionEntity> eventSubscriptions = eventSubscriptionEntityManager.findEventSubscriptionsByExecution(executionEntity.getId());
-    for (EventSubscriptionEntity eventSubscription : eventSubscriptions) {
-      eventSubscriptionEntityManager.delete(eventSubscription);
+    if (!enableExecutionRelationshipCounts 
+        || (enableExecutionRelationshipCounts && ((CountingExecutionEntity) executionEntity).getEventSubscriptionCount() > 0)) {
+      EventSubscriptionEntityManager eventSubscriptionEntityManager = getEventSubscriptionEntityManager();
+      List<EventSubscriptionEntity> eventSubscriptions = eventSubscriptionEntityManager.findEventSubscriptionsByExecution(executionEntity.getId());
+      for (EventSubscriptionEntity eventSubscription : eventSubscriptions) {
+        eventSubscriptionEntityManager.delete(eventSubscription);
+      }
     }
     
   }
